tests/wpt/web-platform-tests/workers/WorkerLocation_search_fragment.htm

<!DOCTYPE html>
<title> WorkerLocation.search with &lt;fragment&gt; in &lt;query&gt; </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/WorkerLocation.js?test#');
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.search, "?test");
  });
});
</script>
